Share code between code runners
