+++ benches/impl_path_string_for_evaluation_context.rs
@@ -1,14 +1,7 @@
-use std::borrow::Cow;
-use std::collections::BTreeMap;
-
-use async_graphql::{Name, Value};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
-use hyper::header::HeaderValue;
-use hyper::HeaderMap;
-use indexmap::IndexMap;
-use once_cell::sync::Lazy;
+use tailcall::benchmark::{assert_test, MockGraphqlContext, TEST_HEADERS, TEST_VARS};
 use tailcall::http::RequestContext;
-use tailcall::lambda::{EvaluationContext, ResolverContextLike};
+use tailcall::lambda::EvaluationContext;
 use tailcall::path_string::PathString;
 
 const INPUT_VALUE: &[&[&str]] = &[
@@ -33,98 +26,8 @@ const HEADERS_VALUE: &[&[&str]] = &[&["headers", "existing"], &["headers", "miss
 
 const VARS_VALUE: &[&[&str]] = &[&["vars", "existing"], &["vars", "missing"]];
 
-static TEST_VALUES: Lazy<Value> = Lazy::new(|| {
-  let mut root = IndexMap::new();
-  let mut nested = IndexMap::new();
-
-  nested.insert(Name::new("existing"), Value::String("nested-test".to_owned()));
-
-  root.insert(Name::new("root"), Value::String("root-test".to_owned()));
-  root.insert(Name::new("nested"), Value::Object(nested));
-
-  Value::Object(root)
-});
-
-static TEST_ARGS: Lazy<IndexMap<Name, Value>> = Lazy::new(|| {
-  let mut root = IndexMap::new();
-  let mut nested = IndexMap::new();
-
-  nested.insert(Name::new("existing"), Value::String("nested-test".to_owned()));
-
-  root.insert(Name::new("root"), Value::String("root-test".to_owned()));
-  root.insert(Name::new("nested"), Value::Object(nested));
-
-  root
-});
-
-static TEST_HEADERS: Lazy<HeaderMap> = Lazy::new(|| {
-  let mut map = HeaderMap::new();
-
-  map.insert("x-existing", HeaderValue::from_static("header"));
-
-  map
-});
-
-static TEST_VARS: Lazy<BTreeMap<String, String>> = Lazy::new(|| {
-  let mut map = BTreeMap::new();
-
-  map.insert("existing".to_owned(), "var".to_owned());
-
-  map
-});
-
 fn to_bench_id(input: &[&str]) -> BenchmarkId {
   BenchmarkId::new("input", input.join("."))
-}
-
-struct MockGraphqlContext;
-
-impl<'a> ResolverContextLike<'a> for MockGraphqlContext {
-  fn value(&'a self) -> Option<&'a Value> {
-    Some(&TEST_VALUES)
-  }
-
-  fn args(&'a self) -> Option<&'a IndexMap<Name, Value>> {
-    Some(&TEST_ARGS)
-  }
-}
-
-// assert that everything was set up correctly for the benchmark
-fn assert_test(eval_ctx: &EvaluationContext<'_, MockGraphqlContext>) {
-  // value
-  assert_eq!(
-    eval_ctx.path_string(&["value", "root"]),
-    Some(Cow::Borrowed("root-test"))
-  );
-  assert_eq!(
-    eval_ctx.path_string(&["value", "nested", "existing"]),
-    Some(Cow::Borrowed("nested-test"))
-  );
-  assert_eq!(eval_ctx.path_string(&["value", "missing"]), None);
-  assert_eq!(eval_ctx.path_string(&["value", "nested", "missing"]), None);
-
-  // args
-  assert_eq!(
-    eval_ctx.path_string(&["args", "root"]),
-    Some(Cow::Borrowed("root-test"))
-  );
-  assert_eq!(
-    eval_ctx.path_string(&["args", "nested", "existing"]),
-    Some(Cow::Borrowed("nested-test"))
-  );
-  assert_eq!(eval_ctx.path_string(&["args", "missing"]), None);
-  assert_eq!(eval_ctx.path_string(&["args", "nested", "missing"]), None);
-
-  // headers
-  assert_eq!(
-    eval_ctx.path_string(&["headers", "x-existing"]),
-    Some(Cow::Borrowed("header"))
-  );
-  assert_eq!(eval_ctx.path_string(&["headers", "x-missing"]), None);
-
-  // vars
-  assert_eq!(eval_ctx.path_string(&["vars", "existing"]), Some(Cow::Borrowed("var")));
-  assert_eq!(eval_ctx.path_string(&["vars", "missing"]), None);
 }
 
 fn bench_main(c: &mut Criterion) {
