+++ benches/request_template_bench.rs
@@ -1,48 +1,13 @@
-use std::borrow::Cow;
-
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use derive_setters::Setters;
-use hyper::HeaderMap;
 use serde_json::json;
-use tailcall::endpoint::Endpoint;
-use tailcall::has_headers::HasHeaders;
-use tailcall::path_string::PathString;
-use tailcall::request_template::RequestTemplate;
+use tailcall::benchmark::{create_request_templates, Context};
 
-#[derive(Setters)]
-struct Context {
-  pub value: serde_json::Value,
-  pub headers: HeaderMap,
-}
-
-impl Default for Context {
-  fn default() -> Self {
-    Self { value: serde_json::Value::Null, headers: HeaderMap::new() }
-  }
-}
-impl PathString for Context {
-  fn path_string<T: AsRef<str>>(&self, parts: &[T]) -> Option<Cow<'_, str>> {
-    self.value.path_string(parts)
-  }
-}
-impl HasHeaders for Context {
-  fn headers(&self) -> &HeaderMap {
-    &self.headers
-  }
-}
 fn benchmark_to_request(c: &mut Criterion) {
-  let tmpl_mustache = RequestTemplate::try_from(Endpoint::new(
-    "http://localhost:3000/{{args.b}}?a={{args.a}}&b={{args.b}}&c={{args.c}}".to_string(),
-  ))
-  .unwrap();
-
-  let tmpl_literal =
-    RequestTemplate::try_from(Endpoint::new("http://localhost:3000/foo?a=bar&b=foo&c=baz".to_string())).unwrap();
-
+  let (tmpl_literal, tmpl_mustache) = create_request_templates();
   let ctx = Context::default().value(json!({
-    "args": {
-      "b": "foo"
-    }
+      "args": {
+          "b": "foo"
+      }
   }));
 
   c.bench_function("with_mustache_literal", |b| {
