+++ src/benchmark/mod.rs
@@ -1,0 +1,9 @@
+pub use self::assert_test::*;
+pub use self::create_request_templates::*;
+pub use self::gather_path_matches::*;
+pub use self::run_data_loader_benchmark::*;
+
+pub mod assert_test;
+pub mod create_request_templates;
+pub mod gather_path_matches;
+pub mod run_data_loader_benchmark;
